virtio: Prepare for 64 feature bits
virtio 1.0 supports an arbitrary number of feature bits. However, so far
no more than 64 are used (compared to 32 in virtio 0.9). Adjust data
types to support 64 feature bits.

Later, we may want to use bitmaps and setbit(), ... to support even more
feature bits.

ok mlarkin@

+++ sys/dev/pv/vioblk.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: vioblk.c,v 1.9 2017/08/10 18:06:58 reyk Exp $	*/
+/*	$OpenBSD: vioblk.c,v 1.10 2019/01/08 16:23:01 sf Exp $	*/
 
 /*
  * Copyright (c) 2012 Stefan Fritsch.
@@ -168,7 +168,7 @@ vioblk_attach(struct device *parent, struct device *self, void *aux)
 	struct vioblk_softc *sc = (struct vioblk_softc *)self;
 	struct virtio_softc *vsc = (struct virtio_softc *)parent;
 	struct scsibus_attach_args saa;
-	uint32_t features;
+	uint64_t features;
 	int qsize;
 
 	vsc->sc_vqs = &sc->sc_vq[0];
